Restructure timer module, split timer types
