+++ src/hyperfine/shell.rs
@@ -1,15 +1,11 @@
+use std;
 use std::io;
 use std::process::{Command, Stdio};
 
-#[cfg(not(windows))]
-use std::process::ExitStatus;
-
-#[cfg(windows)]
-use std::process::Child;
 
 /// Run a standard shell command
 #[cfg(not(windows))]
-pub fn run_shell_command(command: &str) -> io::Result<ExitStatus> {
+pub fn run_shell_command(command: &str) -> io::Result<std::process::ExitStatus> {
     Command::new("sh")
         .arg("-c")
         .arg(command)
@@ -21,7 +17,7 @@ pub fn run_shell_command(command: &str) -> io::Result<ExitStatus> {
 
 /// Run a Windows shell command using cmd.exe
 #[cfg(windows)]
-pub fn run_shell_command(command: &str) -> io::Result<Child> {
+pub fn run_shell_command(command: &str) -> io::Result<std::process::Child> {
     Command::new("cmd")
         .arg("/C")
         .arg(command)
