+++ src/hyperfine/timer/internal.rs
@@ -1,0 +1,19 @@
+use hyperfine::internal::Second;
+
+#[derive(Debug, Copy, Clone)]
+pub struct CPUTimes {
+    /// Total amount of time spent executing in user mode
+    pub user_usec: i64,
+
+    /// Total amount of time spent executing in kernel mode
+    pub system_usec: i64,
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct CPUInterval {
+    /// Total amount of time spent executing in user mode
+    pub user: Second,
+
+    /// Total amount of time spent executing in kernel mode
+    pub system: Second,
+}
