+++ src/hyperfine/timer/mod.rs
@@ -1,5 +1,12 @@
+mod internal;
+
 pub mod wallclocktimer;
-pub mod cputimer;
+
+#[cfg(not(windows))]
+pub mod unix_timer;
+
+#[cfg(windows)]
+pub mod windows_timer;
 
 /// Defines a general timer with start/stop functionality.
 pub trait Timer {
